moved the pressure group calculation class
moved the pressure group calculation class from dive to logic
unsure if the original dive class is referenced because the class in logic now does not import the class in dive...
