+++ src/main/java/seedu/divelog/model/dive/PressureGroup.java
@@ -2,11 +2,6 @@ package seedu.divelog.model.dive;
 
 //@@author arjo129
 
-import org.json.JSONArray;
-import org.json.JSONException;
-
-import seedu.divelog.model.divetables.PadiDiveTable;
-
 /**
  * Represents a pressure group
  */
@@ -27,25 +22,8 @@ public class PressureGroup {
         this.pressureGroup = pressureGroup.toUpperCase();
     }
 
-    //@@author shuanang
-    public PressureGroup(String pressureGroup, int newDepth, int minutesRepeatDive) throws JSONException {
-        assert pressureGroup.length() == 1;
-        assert Character.isLetter(pressureGroup.charAt(0));
-        this.pressureGroup = pressureGroup.toUpperCase();
-        PadiDiveTable padiDiveTable = PadiDiveTable.getInstance();
-        JSONArray arr = padiDiveTable.depthToTimes(new DepthProfile(newDepth), new PressureGroup(pressureGroup));
-        assert minutesRepeatDive <= arr.getInt(1);
-        this.totalBottomTime = arr.getInt(0) + minutesRepeatDive;
-        PressureGroup pg = padiDiveTable.depthToPg(new DepthProfile(newDepth), this.totalBottomTime);
-        newPg = pg.getPressureGroup();
-    }
-
     public String getPressureGroup() {
         return pressureGroup;
-    }
-
-    public String getNewPressureGroup() {
-        return newPg;
     }
 
     /**
