+++ src/test/java/seedu/divelog/logic/pressuregroup/PressureGroupTest.java
@@ -1,16 +1,17 @@
-package seedu.divelog.model.dive;
+package seedu.divelog.logic.pressuregroup;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.json.JSONException;
 import org.junit.Test;
 
+
 //@@author shuanang
 public class PressureGroupTest {
 
     @Test
     public void pressureGroup_repeatDive() throws JSONException {
-        PressureGroup pg = new PressureGroup("d", 14, 50);
+        seedu.divelog.logic.pressuregroup.PressureGroup pg = new PressureGroup("d", 14, 50);
         assertEquals(pg.getNewPressureGroup(), "T");
     }
 }
